Factor `DataFeed` client API into `brokers.data`
